Further clean up Cell by moving cell content rendering to CellContent
